requirements.txt
PySide6>=6.5
openpyxl>=3.1
customtkinter>=5.2.0
Pillow>=10.0.0
reportlab>=4.0.0
